requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
dill==0.3.8
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.4.3
multiprocess==0.70.16
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
termcolor==3.0.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
